requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.4
pyjwt==2.8.0
werkzeug==3.0.1
scikit-learn==1.3.2